pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Anderson_Carvalho <andersonlimacrv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
notion-client = "^2.2.1"
python-dotenv = "^1.0.0"
pywhatkit = "^5.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
